fix: Round 17 — CancellationException guards, customReason cap, .toDouble() removed, truck plural, currency locale, My Trips string

+++ app/src/main/java/com/weelo/logistics/presentation/booking/BookingTrackingActivity.kt
@@ -616,8 +616,8 @@ class BookingTrackingActivity : AppCompatActivity(), OnMapReadyCallback {
                 trackingData.forEach { truck ->
                     addOrUpdateDriverMarker(
                         tripId = truck.tripId,
-                        lat = truck.latitude.toDouble(),
-                        lng = truck.longitude.toDouble(),
+                        lat = truck.latitude,
+                        lng = truck.longitude,
                         status = truck.status,
                         title = truck.vehicleNumber
                     )
